backend/requirements.txt
# RAG Chatbot Backend Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI/ML
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.6
langchain-core>=0.1.0

# Vector Store
faiss-cpu>=1.7.4

# PDF Processing
pypdf>=3.17.0

# Database (async)
aiosqlite>=0.19.0

# Environment
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.5.0

# Logging (Tier 3)
structlog>=24.0.0

# Testing (Tier 3 - dev dependencies)
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

